+++ src/textual/dom.py
@@ -135,12 +135,18 @@ class DOMNode(MessagePump):
     # Virtual DOM nodes
     COMPONENT_CLASSES: ClassVar[set[str]] = set()
 
+    BINDING_GROUP_TITLE: str | None = None
+    """Title of widget used where bindings are displayed (such as in the key panel)."""
+
     # Mapping of key bindings
     BINDINGS: ClassVar[list[BindingType]] = []
 
     # Indicates if the CSS should be automatically scoped
     SCOPED_CSS: ClassVar[bool] = True
     """Should default css be limited to the widget type?"""
+
+    HELP: ClassVar[str | None] = None
+    """Optional help text shown in help panel (Markdown format)."""
 
     # True if this node inherits the CSS from the base class.
     _inherit_css: ClassVar[bool] = True
@@ -206,9 +212,7 @@ class DOMNode(MessagePump):
         self._reactive_connect: (
             dict[str, tuple[MessagePump, Reactive | object]] | None
         ) = None
-
         self._pruning = False
-
         super().__init__()
 
     def set_reactive(
@@ -282,7 +286,6 @@ class DOMNode(MessagePump):
                 yield WorldClock("Europe/Paris").data_bind(WorldClockApp.time)
                 yield WorldClock("Asia/Tokyo").data_bind(WorldClockApp.time)
             ```
-
 
         Raises:
             ReactiveError: If the data wasn't bound.
@@ -1169,13 +1172,10 @@ class DOMNode(MessagePump):
         """Watches for modifications to reactive attributes on another object.
 
         Example:
-
-            Here's how you could detect when the app changes from dark to light mode (and vice versa).
-
             ```python
-            def on_dark_change(old_value:bool, new_value:bool):
+            def on_dark_change(old_value:bool, new_value:bool) -> None:
                 # Called when app.dark changes.
-                print("App.dark when from {old_value} to {new_value}")
+                print("App.dark went from {old_value} to {new_value}")
 
             self.watch(self.app, "dark", self.on_dark_change, init=False)
             ```