+++ src/textual/widget.py
@@ -304,6 +304,15 @@ class Widget(DOMNode):
     BORDER_SUBTITLE: ClassVar[str] = ""
     """Initial value for border_subtitle attribute."""
 
+    ALLOW_MAXIMIZE: ClassVar[bool | None] = None
+    """Defines default logic to allow the widget to be maximized.
+    
+    - `None` Use default behavior (Focusable widgets may be maximized)
+    - `False` Do not allow widget to be maximized
+    - `True` Allow widget to be maximized
+    
+    """
+
     can_focus: bool = False
     """Widget may receive focus."""
     can_focus_children: bool = True
@@ -515,6 +524,15 @@ class Widget(DOMNode):
         )
 
     @property
+    def allow_maximize(self) -> bool:
+        """Check if the widget may be maximized.
+
+        Returns:
+            `True` if the widget may be maximized, or `False` if it should not be maximized.
+        """
+        return self.can_focus if self.ALLOW_MAXIMIZE is None else self.ALLOW_MAXIMIZE
+
+    @property
     def offset(self) -> Offset:
         """Widget offset from origin.
 
@@ -557,6 +575,14 @@ class Widget(DOMNode):
             if widget is self:
                 return True
         return False
+
+    @property
+    def is_maximized(self) -> bool:
+        """Is this widget maximized?"""
+        try:
+            return self.screen.maximized is self
+        except NoScreen:
+            return False
 
     def anchor(self, *, animate: bool = False) -> None:
         """Anchor the widget, which scrolls it into view (like [scroll_visible][textual.widget.Widget.scroll_visible]),
@@ -1233,6 +1259,7 @@ class Widget(DOMNode):
             content_width = Fraction(content_container.width - margin.width)
         elif is_auto_width:
             # When width is auto, we want enough space to always fit the content
+
             content_width = Fraction(
                 self.get_content_width(content_container - margin.totals, viewport)
             )
@@ -1348,7 +1375,6 @@ class Widget(DOMNode):
         """
 
         if self.is_container:
-            assert self._layout is not None
             width = self._layout.get_content_width(self, container, viewport)
             return width
 
@@ -1368,6 +1394,7 @@ class Widget(DOMNode):
             width = min(width, container.width)
 
         self._content_width_cache = (cache_key, width)
+
         return width
 
     def get_content_height(self, container: Size, viewport: Size, width: int) -> int:
@@ -1997,7 +2024,6 @@ class Widget(DOMNode):
                 break
             if node.styles.has_rule("layers"):
                 layers = node.styles.layers
-
         return layers
 
     @property
@@ -2647,7 +2673,7 @@ class Widget(DOMNode):
             level: Minimum level required for the animation to take place (inclusive).
         """
         self.scroll_to(
-            y=self.scroll_y - self.container_size.height,
+            y=self.scroll_y - self.scrollable_content_region.height,
             animate=animate,
             speed=speed,
             duration=duration,
@@ -2680,7 +2706,7 @@ class Widget(DOMNode):
             level: Minimum level required for the animation to take place (inclusive).
         """
         self.scroll_to(
-            y=self.scroll_y + self.container_size.height,
+            y=self.scroll_y + self.scrollable_content_region.height,
             animate=animate,
             speed=speed,
             duration=duration,
@@ -2715,7 +2741,7 @@ class Widget(DOMNode):
         if speed is None and duration is None:
             duration = 0.3
         self.scroll_to(
-            x=self.scroll_x - self.container_size.width,
+            x=self.scroll_x - self.scrollable_content_region.width,
             animate=animate,
             speed=speed,
             duration=duration,
@@ -2750,7 +2776,7 @@ class Widget(DOMNode):
         if speed is None and duration is None:
             duration = 0.3
         self.scroll_to(
-            x=self.scroll_x + self.container_size.width,
+            x=self.scroll_x + self.scrollable_content_region.width,
             animate=animate,
             speed=speed,
             duration=duration,
@@ -3045,7 +3071,7 @@ class Widget(DOMNode):
         name = cls.__name__
         if not name[0].isupper() and not name.startswith("_"):
             raise BadWidgetName(
-                f"Widget subclass {name!r} should be capitalised or start with '_'."
+                f"Widget subclass {name!r} should be capitalized or start with '_'."
             )
 
         super().__init_subclass__(