+++ tests/snapshot_tests/test_snapshots.py
@@ -3,10 +3,11 @@ from pathlib import Path
 import pytest
 
 from tests.snapshot_tests.language_snippets import SNIPPETS
+from textual.app import App, ComposeResult
+from textual.containers import Vertical
 from textual.pilot import Pilot
-from textual.widgets.text_area import Selection, BUILTIN_LANGUAGES
-from textual.widgets import RichLog, TextArea, Input, Button
-from textual.widgets.text_area import TextAreaTheme
+from textual.widgets import Button, Input, RichLog, TextArea, Footer
+from textual.widgets.text_area import BUILTIN_LANGUAGES, Selection, TextAreaTheme
 
 # These paths should be relative to THIS directory.
 WIDGET_EXAMPLES_DIR = Path("../../docs/examples/widgets")
@@ -805,7 +806,7 @@ def test_tooltips_in_compound_widgets(snap_compare):
     async def run_before(pilot) -> None:
         await pilot.pause()
         await pilot.hover("ProgressBar")
-        await pilot.pause(0.3)
+        await pilot.pause(0.4)
         await pilot.pause()
 
     assert snap_compare(SNAPSHOT_APPS_DIR / "tooltips.py", run_before=run_before)
@@ -1204,7 +1205,7 @@ def test_example_color_command(snap_compare):
     """Test the color_command example."""
     assert snap_compare(
         EXAMPLES_DIR / "color_command.py",
-        press=["ctrl+backslash", "r", "e", "d", "down", "enter"],
+        press=[App.COMMAND_PALETTE_BINDING, "r", "e", "d", "down", "enter"],
     )
 
 
@@ -1419,3 +1420,119 @@ def test_auto_height_scrollbar(snap_compare):
 def test_bind_override(snap_compare):
     """Regression test for https://github.com/Textualize/textual/issues/4755"""
     assert snap_compare(SNAPSHOT_APPS_DIR / "bind_override.py")
+
+
+def test_command_palette_dismiss_escape(snap_compare):
+    """Regression test for https://github.com/Textualize/textual/issues/4856"""
+
+    async def run_before(pilot: Pilot):
+        await pilot.press(App.COMMAND_PALETTE_BINDING)
+        await pilot.pause()
+        await pilot.press("escape")
+
+    assert snap_compare(
+        SNAPSHOT_APPS_DIR / "command_palette_dismiss.py", run_before=run_before
+    )
+
+
+def test_command_palette_dismiss_escape_no_results(snap_compare):
+    """Regression test for https://github.com/Textualize/textual/issues/4856"""
+
+    async def run_before(pilot: Pilot):
+        await pilot.press(App.COMMAND_PALETTE_BINDING)
+        await pilot.pause()
+        await pilot.press(*"foo")
+        await pilot.app.workers.wait_for_complete()
+        await pilot.press("escape")
+
+    assert snap_compare(
+        SNAPSHOT_APPS_DIR / "command_palette_dismiss.py", run_before=run_before
+    )
+
+
+def test_command_palette_key_change(snap_compare):
+    """Regression test for https://github.com/Textualize/textual/issues/4887"""
+    assert snap_compare(SNAPSHOT_APPS_DIR / "command_palette_key.py")
+
+
+def test_split(snap_compare):
+    """Test split rule."""
+    assert snap_compare(SNAPSHOT_APPS_DIR / "split.py", terminal_size=(100, 30))
+
+
+def test_system_commands(snap_compare):
+    """The system commands should appear in the command palette."""
+
+    class SimpleApp(App):
+        def compose(self) -> ComposeResult:
+            yield Input()
+
+    assert snap_compare(
+        SimpleApp(),
+        terminal_size=(100, 30),
+        press=["ctrl+p"],
+    )
+
+
+def test_help_panel(snap_compare):
+    """Test help panel."""
+
+    class HelpPanelApp(App):
+        def compose(self) -> ComposeResult:
+            yield Input()
+
+    async def run_before(pilot: Pilot):
+        pilot.app.query(Input).first().cursor_blink = False
+        await pilot.press(App.COMMAND_PALETTE_BINDING)
+        await pilot.pause()
+        await pilot.press(*"keys")
+        await pilot.press("enter")
+        await pilot.app.workers.wait_for_complete()
+
+    assert snap_compare(HelpPanelApp(), terminal_size=(100, 30), run_before=run_before)
+
+
+def test_scroll_page_down(snap_compare):
+    """Regression test for https://github.com/Textualize/textual/issues/4914"""
+    # Should show 25 at the top
+    assert snap_compare(
+        SNAPSHOT_APPS_DIR / "scroll_page.py", press=["pagedown"], terminal_size=(80, 25)
+    )
+
+
+def test_maximize(snap_compare):
+    """Check that maximize isolates a single widget."""
+
+    class MaximizeApp(App):
+        BINDINGS = [("m", "screen.maximize", "maximize focused widget")]
+
+        def compose(self) -> ComposeResult:
+            yield Button("Hello")
+            yield Button("World")
+            yield Footer()
+
+    assert snap_compare(MaximizeApp(), press=["m"])
+
+
+def test_maximize_container(snap_compare):
+    """Check maximizing a widget in a maximizeable container, maximizes the container."""
+
+    class FormContainer(Vertical):
+        ALLOW_MAXIMIZE = True
+        DEFAULT_CSS = """
+        FormContainer {
+            width: 50%;
+            border: blue;            
+        }
+        """
+
+    class MaximizeApp(App):
+        BINDINGS = [("m", "screen.maximize", "maximize focused widget")]
+
+        def compose(self) -> ComposeResult:
+            with FormContainer():
+                yield Button("Hello")
+                yield Button("World")
+            yield Footer()
+
+    assert snap_compare(MaximizeApp(), press=["m"])