+++ tests/test_screens.py
@@ -40,7 +40,7 @@ async def test_installed_screens():
     class ScreensApp(App):
         SCREENS = {
             "home": Screen,  # Screen type
-            "one": Screen(),  # Screen instance
+            "one": Screen,  # Screen instance, disallowed as of #4893
             "two": Screen,  # Callable[[], Screen]
         }
 
@@ -354,7 +354,7 @@ async def test_switch_screen_no_op():
         pass
 
     class MyApp(App[None]):
-        SCREENS = {"screen": MyScreen()}
+        SCREENS = {"screen": MyScreen}
 
         def on_mount(self):
             self.push_screen("screen")
@@ -379,8 +379,8 @@ async def test_switch_screen_updates_results_callback_stack():
 
     class MyApp(App[None]):
         SCREENS = {
-            "a": ScreenA(),
-            "b": ScreenB(),
+            "a": ScreenA,
+            "b": ScreenB,
         }
 
         def callback(self, _):
@@ -407,7 +407,7 @@ async def test_screen_receives_mouse_move_events():
             MouseMoveRecordingScreen.mouse_events.append(event)
 
     class SimpleApp(App[None]):
-        SCREENS = {"a": MouseMoveRecordingScreen()}
+        SCREENS = {"a": MouseMoveRecordingScreen}
 
         def on_mount(self):
             self.push_screen("a")
@@ -439,7 +439,7 @@ async def test_mouse_move_event_bubbles_to_screen_from_widget():
             MouseMoveRecordingScreen.mouse_events.append(event)
 
     class SimpleApp(App[None]):
-        SCREENS = {"a": MouseMoveRecordingScreen()}
+        SCREENS = {"a": MouseMoveRecordingScreen}
 
         def on_mount(self):
             self.push_screen("a")
@@ -537,3 +537,91 @@ async def test_default_custom_screen() -> None:
         assert len(app.screen_stack) == 1
         assert isinstance(app.screen_stack[0], CustomScreen)
         assert app.screen is app.screen_stack[0]
+
+
+async def test_disallow_screen_instances() -> None:
+    """Test that screen instances are disallowed."""
+
+    class CustomScreen(Screen):
+        pass
+
+    with pytest.raises(ValueError):
+
+        class Bad(App):
+            SCREENS = {"a": CustomScreen()}  # type: ignore
+
+    with pytest.raises(ValueError):
+
+        class Worse(App):
+            MODES = {"a": CustomScreen()}  # type: ignore
+
+    # While we're here, let's make sure that other types
+    # are disallowed.
+    with pytest.raises(TypeError):
+
+        class Terrible(App):
+            MODES = {"a": 42, "b": CustomScreen}  # type: ignore
+
+    with pytest.raises(TypeError):
+
+        class Worst(App):
+            MODES = {"OK": CustomScreen, 1: 2}  # type: ignore
+
+
+async def test_worker_cancellation():
+    """Regression test for https://github.com/Textualize/textual/issues/4884
+
+    The MRE below was pushing a screen in an exclusive worker.
+    This was previously breaking because the second time the worker was launched,
+    it cancelled the first one which was awaiting the screen.
+
+    """
+    from textual import on, work
+    from textual.app import App
+    from textual.containers import Vertical
+    from textual.screen import Screen
+    from textual.widgets import Button, Footer, Label
+
+    class InfoScreen(Screen[bool]):
+        def __init__(self, question: str) -> None:
+            self.question = question
+            super().__init__()
+
+        def compose(self) -> ComposeResult:
+            yield Vertical(
+                Label(self.question, id="info-label"),
+                Button("Ok", variant="primary", id="ok"),
+                id="info-vertical",
+            )
+            yield Footer()
+
+        @on(Button.Pressed, "#ok")
+        def handle_ok(self) -> None:
+            self.dismiss(True)  # Changed the `dismiss` result to compatible type
+
+    class ExampleApp(App):
+        BINDINGS = [("i", "info", "Info")]
+
+        screen_count = 0
+
+        def compose(self) -> ComposeResult:
+            yield Label("This is the default screen")
+            yield Footer()
+
+        @work(exclusive=True)
+        async def action_info(self) -> None:
+            # Since this is an exclusive worker, the second time it is called,
+            # the original `push_screen_wait` is also cancelled
+            self.screen_count += 1
+            await self.push_screen_wait(
+                InfoScreen(f"This is info screen #{self.screen_count}")
+            )
+
+    app = ExampleApp()
+    async with app.run_test() as pilot:
+        # Press i twice to launch 2 InfoScreens
+        await pilot.press("i")
+        await pilot.press("i")
+        # Press enter to activate button to dismiss them
+        await pilot.press("enter")
+        await pilot.press("enter")